+++ src/devices/index.ts
@@ -269,6 +269,7 @@ import {definitions as siglis} from "./siglis";
 import {definitions as sikom} from "./sikom";
 import {definitions as siliconLabs} from "./silicon_labs";
 import {definitions as simon} from "./simon";
+import {definitions as simplaHome} from "./simpla_home";
 import {definitions as sinope} from "./sinope";
 import {definitions as siterwell} from "./siterwell";
 import {definitions as skydance} from "./skydance";
@@ -612,6 +613,7 @@ const definitions: DefinitionWithExtend[] = [
     ...siliconLabs,
     ...sikom,
     ...simon,
+    ...simplaHome,
     ...sinope,
     ...siterwell,
     ...skydance,